requirements.txt
paho_mqtt==1.5.1
python-dotenv==0.17.1
PyYAML==5.4.1
